cv_service/requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
python-multipart==0.0.6

# Vector Database
pinecone==7.3.0

# Computer Vision & ML
opencv-python==4.8.1.78
scikit-learn==1.3.2
numpy==1.24.3
pillow==10.0.1
scipy==1.11.4

# CV Pipeline - Object Detection
ultralytics>=8.0.0         # YOLOv8

# CV Pipeline - Deep Learning (CPU-only for faster installation)
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
torchvision>=0.15.0
